requirements.txt
Flask==3.0.0
flask-cors==6.0.1
requests==2.31.0
python-whatsapp-bot==2.0.3
qrcode==8.2
Pillow==10.1.0


qrcode
Pillow

